Add autocommit
